requirements.txt
pyyaml>=6.0.2
docker>=7.1.0
requests>=2.26.0
cryptography>=41.0.0
pyjwt>=2.8.0
watchdog>=3.0.0
